feat(ui): separate kapacitor flux tasks API
